Split ozone-typescript-client index into files Implement roleClient
