+++ packages/ozone-typescript-client/src/ozoneClient/Credentials.ts
@@ -1,0 +1,10 @@
+
+export interface AuthInfo {
+	principalClass: string,
+	principalId: string,
+	sessionId: string,
+}
+
+export abstract class OzoneCredentials {
+	abstract authenticate(ozoneURL: string): Promise<AuthInfo>
+}
