+++ packages/ozone-typescript-client/src/ozoneClient/clientConfiguration.ts
@@ -1,0 +1,9 @@
+import {OzoneCredentials} from "./Credentials";
+
+export interface ClientConfiguration {
+	ozoneURL: string
+	ozoneInstanceId?: string
+	ozoneCredentials?: OzoneCredentials
+	webSocketsURL?: string
+	defaultTimeout?: number
+}
